Improve support for Type 5 high density Tagi (>8K).
Add extended commands as defined by NFC Forum for Type 5.
This set of commands allows to address range of blocks above 8K.
NDEF Native support for high density STM25DV tags is improved.
Large NDEFs (VCARD with a photo) are now supported.

+++ src/nfc/include/rw_api.h
@@ -1104,7 +1104,7 @@ extern tNFC_STATUS RW_I93ReadMultipleBlocks(uint16_t first_block_number,
 **                  NFC_STATUS_FAILED if other error
 **
 *******************************************************************************/
-extern tNFC_STATUS RW_I93WriteMultipleBlocks(uint8_t first_block_number,
+extern tNFC_STATUS RW_I93WriteMultipleBlocks(uint16_t first_block_number,
                                              uint16_t number_blocks,
                                              uint8_t* p_data);
